refactor: Used shared registry types in block streamer
