+++ block-streamer/src/rules/outcomes_reducer.rs
@@ -115,9 +115,10 @@ fn build_indexer_rule_match_payload(
 
 #[cfg(test)]
 mod tests {
+    use registry_types::{IndexerRule, IndexerRuleKind, MatchingRule, Status};
+
     use crate::rules::outcomes_reducer::reduce_indexer_rule_matches_from_outcomes;
     use crate::rules::types::{ChainId, IndexerRuleMatch};
-    use crate::rules::{IndexerRule, IndexerRuleKind, MatchingRule, Status};
 
     #[tokio::test]
     async fn match_wildcard_no_match() {
